+++ crates/amaru-bootstrap/src/cardano_node/tvar.rs
@@ -248,7 +248,7 @@ fn decode_tvar_entry(input: &[u8], output: &[u8]) -> anyhow::Result<(Transaction
 
     let input = TransactionInput {
         transaction_id: Hash::from(&input[..32]),
-        index: u16::from_be_bytes([input[32], input[33]]).into(),
+        index: u16::from_be_bytes([input[32], input[33]]),
     };
     let output = mempack::decode_transaction_output(output)?;