+++ crates/amaru-ledger/src/rules/transaction/phase_one/outputs.rs
@@ -58,6 +58,7 @@ pub enum SupplementalDatumPolicy {
     Disallow,
 }
 
+#[expect(clippy::too_many_arguments)]
 pub fn execute<C>(
     context: &mut C,
     arena_pool: &ArenaPool,
@@ -65,7 +66,8 @@ pub fn execute<C>(
     network: Network,
     outputs: Vec<MemoizedTransactionOutput>,
     supplemental_datum_policy: SupplementalDatumPolicy,
-    construct_utxo: impl Fn(&mut C, u64, &Value) -> Option<TransactionInput>,
+    first_index: u16,
+    construct_utxo: impl Fn(&mut C, u16, &Value) -> Option<TransactionInput>,
 ) -> Result<(), InvalidOutputs>
 where
     C: WitnessSlice + UtxoSlice + BalanceSlice,
@@ -94,7 +96,8 @@ where
                 .unwrap_or_else(|element| invalid_outputs.push(WithPosition { position, element }));
         }
 
-        if let Some(input) = construct_utxo(context, position as u64, &output.value) {
+        let index = first_index + position as u16;
+        if let Some(input) = construct_utxo(context, index, &output.value) {
             context.produce(input, output);
         }
     }